Create README.md
Getting and Cleaning Data Week 4 Project
Repo for Coursera's "Getting and Cleaning Data" Week 4 project. Contains R code, codebook, and tidy data set for the assignment.

Files:
CodeBook.md describes the how to use all this, variables, the data, and any transformations or work that was performed to clean up the data.

run_analysis.R contains all the code to perform the analyses described in the 5 steps. They can be launched in RStudio by just importing the file. Here are the five steps:

1. Merges the training and the test sets to create one data set.
2. Extracts only the measurements on the mean and standard deviation for each measurement.
3. Uses descriptive activity names to name the activities in the data set
4. Appropriately labels the data set with descriptive variable names.
5. From the data set in step 4, creates a second, independent tidy data set with the average of each variable for each activity and each subject.

tidyData.txt is the output of the final step